cloud_functions/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.3
matplotlib==3.10.0
functions-framework==3.*
datetime==5.5
requests==2.32.3
cloud-sql-python-connector==1.16.0
gunicorn==20.1.0
# psycopg2==2.9.10
pg8000==1.31.2 